--story=1003277 --user=赖文韬 保存流程评分设置bug https://www.tapd.cn/54247054/s/1086107

+++ src/main/java/codedriver/module/process/service/ProcessServiceImpl.java
@@ -143,11 +143,9 @@ public class ProcessServiceImpl implements ProcessService {
 		}
 
 		/** 保存评分设置 */
-		if(processVo.getProcessScoreTemplateVo() != null){
-		    if(processVo.getProcessScoreTemplateVo().getIsActive() != null && processVo.getProcessScoreTemplateVo().getIsActive().intValue() == 1){
-                scoreTemplateMapper.deleteProcessScoreTemplateByProcessUuid(processVo.getUuid());
-                scoreTemplateMapper.insertProcessScoreTemplate(processVo.getProcessScoreTemplateVo());
-            }
+		if(processVo.getProcessScoreTemplateVo() != null && processVo.getProcessScoreTemplateVo().getIsActive() != null && processVo.getProcessScoreTemplateVo().getIsActive().intValue() == 1){
+			scoreTemplateMapper.deleteProcessScoreTemplateByProcessUuid(processVo.getUuid());
+			scoreTemplateMapper.insertProcessScoreTemplate(processVo.getProcessScoreTemplateVo());
 		}
 
 		return 1;
